Added model loading. We have the viking room

+++ src/window_manager/vulkan/vulkan_buffers.cpp
@@ -200,7 +200,7 @@ void create_vertex_buffer( std::vector<Vertex> vertices, std::shared_ptr<VulkanD
   // vkFreeMemory( vulkanDevice->device, stagingBufferMemory, nullptr );
 }
 
-void create_index_buffer( std::vector<uint16_t> indices, std::shared_ptr<VulkanDevice> vulkanDevice,
+void create_index_buffer( std::vector<uint32_t> indices, std::shared_ptr<VulkanDevice> vulkanDevice,
                           std::shared_ptr<Buffer> indexBuffer,
                           vk::raii::CommandPool &commandPool ) {
   Logger::log( "Creating index buffer...", Logger::DEBUG );
